chore: update project structure
- Organize components according to their type
- Rename components to start with uppercase
