+++ src/components/hero/CTA.tsx
@@ -1,12 +1,11 @@
 "use client";
 
 import Link from "next/link";
-import { buttonVariants } from "./ui/button";
+import { buttonVariants } from "../ui/button";
 import { useEffect, useContext } from "react";
 import anime from "animejs";
 import ScrollContext from "@/context/scrollContext";
-import { Button } from "./ui/button";
-import { TextGenerateEffect } from "./ui/text-generate-effect";
+import { Button } from "../ui/button";
 
 /**
  * Call to action component.
@@ -15,9 +14,12 @@ import { TextGenerateEffect } from "./ui/text-generate-effect";
  *
  * @returns A call to action component that has to be rendered in the hero section
  */
-const CTA = () => {
+export const CTA = () => {
   const targetRef = useContext(ScrollContext);
 
+  /**
+   * Scrolls to the features section when the user clicks on the explore button
+   */
   const scrollToFeatures = () => {
     if (targetRef.current) {
       targetRef.current.scrollIntoView({
@@ -27,13 +29,20 @@ const CTA = () => {
     }
   };
 
-  useEffect(() => {
+  /**
+   * Adds a fade-in effect to the call to action component
+   */
+  const animateCTA = () => {
     anime.timeline({ easing: "linear" }).add({
       targets: ".cta",
       opacity: [0, 1],
       duration: 2000,
       delay: 3000,
     });
+  };
+
+  useEffect(() => {
+    animateCTA();
   }, []);
 
   return (
@@ -50,25 +59,5 @@ const CTA = () => {
         Explore
       </Button>
     </div>
-  );
-};
-
-/**
- * Hero section component. Renders an animated message and a call to action component.
- *
- * @returns A hero section that contains a message and a call to action component
- */
-export const Hero = () => {
-  return (
-    <>
-      <section className="h-[75vh] grid justify-center place-content-center">
-        <article className="space-y-3 max-w-2xl">
-          <header className="space-y-2 font-bold text-7xl">
-            <TextGenerateEffect words="Unleash ideas Embrace innovation" />
-          </header>
-          <CTA />
-        </article>
-      </section>
-    </>
   );
 };
