+++ src/components/navbar.tsx
@@ -1,5 +1,4 @@
 import Link from "next/link";
-import { Button } from "@/components/ui/button";
 
 export const Navbar = () => {
   return (
